Add enum constraint to media_type columns
Media types are extremely important for consistency and integrity, as we rely on them to distinguish the type of each object, so we should enforce an enumeration for them. We can add more types in future, but we definitely want to know if someone tries to push an unknown media type before allowing it.