Update manifests entity and its dependent entities
Manifest records are no longer deduplicated and shared across
repositories but rather scoped by repository. This means that each
`manifests` row will now have a `repository_id` FK for `repositories`.
As a consequence, the link between `manifests` and its dependent
entities, namely `manifest_references`, `layers` and `tags` becomes
broken and needs to be un-deduplicated as well. For this reason we had
to perform all this changes in this single MR, otherwise all tests would
be broken.