Avoid preceding select when deleting blob links
Additionally, this MR uses a digest as key for blob related methods
whenever possible, instead of requiring a models.Blob struct. The
reasoning is that with the partitioned schema, the blobs table no longer
have an `id`, and the primary key now is the digest. Therefore, we can
unequivocally refer to blobs by their digest. Furthermore, when calling
these methods we always have the blob digest, but we may not have a
models.Blob struct, and building one with just the digest to pass it as
argument to these functions is cumbersome.