fix(handlers): always add the migration_path label to HTTP metrics
When `migration.enabled` was `false`, the `migration_path` label would
not be added to the HTTP metrics, but Prometheus expects it to be always
present (either set or not). This was breaking the metrics emission.

This change makes sure that the label is always present. When
`migration.enabled` is `false`, the label is set to an empty string.
When `migration.enabled` is `true`, the label is set to either `old` or
`new`.