fix(api/errcode): extract enclosed error from a storage driver catch-all error
Whenever an unexpected error happens, the storage driver base client
will wrap it in a catch-all error, `storagedriver.Error`. Because of it,
if we get one of these errors we must extract the enclosed error before
trying to identify the best HTTP status for it, otherwise this will
always lead to a 500 Internal Server Error.