fix(handlers): ignore tag not found errors when deleting a manifest
When deleting a manifest through the API, we 1) delete the manifest, 2)
obtain the list of tags (if any) that point to that manifest and 3)
delete them in a loop. If for some reason a tag is deleted between 2 and
3, the request will be halted with a 500 Internal Server error. This is
because we are not ignoring "path not found" errors during 3. This
change fixes that.